Print all possible combinationstwo numbers using putchar commit

+++ 0x01-variables_if_else_while/100-print_comb3.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	int n = '10'
+	long int n = '10'
 	
 	while (n <= '99')
 	{
